refactor: 파일명 규칙 통일 - PascalCase를 kebab-case로 변경
- Chat 컴포넌트 파일명 변경
  - ChatHeader.tsx → chat-header.tsx
  - ChatInput.tsx → chat-input.tsx
  - ChatMessage.tsx → chat-message.tsx
  - ChatMessages.tsx → chat-messages.tsx
  - SummaryButton.tsx → summary-button.tsx

- Chat Hooks 파일명 변경
  - useMentionSystem.ts → use-mention-system.ts
  - useMisoStreaming.ts → use-miso-streaming.ts

- Project 컴포넌트 파일명 변경
  - ProjectCard.tsx → project-card.tsx
  - CreateProjectDialog.tsx → create-project-dialog.tsx
  - ProjectHeader.tsx → project-header.tsx
  - ProjectGrid.tsx → project-grid.tsx
  - ProjectSkeleton.tsx → project-skeleton.tsx
  - ProjectSearchBar.tsx → project-search-bar.tsx

- 모든 관련 import문 업데이트
