+++ components/chat/chat-interface.tsx
@@ -3,11 +3,11 @@
 import { useRef, useEffect, useCallback, useState } from "react"
 import type { Message } from "ai/react"
 import { SummaryModal } from "@/components/chat/summary"
-import { ChatMessages } from "./components/ChatMessages"
-import { ChatInput } from "./components/ChatInput"
-import { SummaryButton } from "./components/SummaryButton"
-import { useMisoStreaming } from "./hooks/useMisoStreaming"
-import { useMentionSystem } from "./hooks/useMentionSystem"
+import { ChatMessages } from "./components/chat-messages"
+import { ChatInput } from "./components/chat-input"
+import { SummaryButton } from "./components/summary-button"
+import { useMisoStreaming } from "./hooks/use-miso-streaming"
+import { useMentionSystem } from "./hooks/use-mention-system"
 import { chatStyles } from "./styles"
 import { ChatInterfaceProps, ExtendedMessage, SummaryData } from "./types"
 import { MentionData } from "@/lib/utils/mention"
