+++ components/chat/components/chat-message.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { motion } from "framer-motion"
 import { Copy, MessageSquareMore } from "lucide-react"
 import { MessageContent } from "@/components/chat/message-content"
-import { ChatHeader } from "./ChatHeader"
+import { ChatHeader } from "./chat-header"
 import { ExtendedMessage } from "../types"
 import type { Message } from "ai/react"
