+++ components/chat/hooks/use-mention-system.ts
